neutron_sync_mode not honored when syncing ACLs
The OVN neutron_sync_mode configuration option is not honored
when syncing ACLs. This problem impacts both service startup
and the neutron_ovn_db_sync command-line utility.

When neutron_sync_mode is set to 'log', ACLs will be synced.
The syncing should only occur when neutron_sync_mode is set
to 'repair'.

Closes-Bug: #1586053

+++ networking_ovn/ovn_nb_sync.py
@@ -227,14 +227,18 @@ class OvnNbSynchronizer(object):
                   (len(list(itertools.chain(*six.itervalues(neutron_acls)))),
                    len(list(itertools.chain(*six.itervalues(nb_acls))))))
 
-        LOG.debug('ACL-SYNC: transaction started @ %s' % str(datetime.now()))
-        with self.ovn_api.transaction(check_error=True) as txn:
-            for acla in list(itertools.chain(*six.itervalues(neutron_acls))):
-                txn.add(self.ovn_api.add_acl(**acla))
-            for aclr in list(itertools.chain(*six.itervalues(nb_acls))):
-                txn.add(self.ovn_api.delete_acl(aclr['lswitch'],
-                                                aclr['lport']))
-        LOG.debug('ACL-SYNC: transaction finished @ %s' % str(datetime.now()))
+        if self.mode == SYNC_MODE_REPAIR:
+            LOG.debug('ACL-SYNC: transaction started @ %s' %
+                      str(datetime.now()))
+            with self.ovn_api.transaction(check_error=True) as txn:
+                for acla in list(itertools.chain(
+                                 *six.itervalues(neutron_acls))):
+                    txn.add(self.ovn_api.add_acl(**acla))
+                for aclr in list(itertools.chain(*six.itervalues(nb_acls))):
+                    txn.add(self.ovn_api.delete_acl(aclr['lswitch'],
+                                                    aclr['lport']))
+            LOG.debug('ACL-SYNC: transaction finished @ %s' %
+                      str(datetime.now()))
 
     def sync_routers_and_rports(self, ctx):
         """Sync Routers between neutron and NB.
